Partially implement Pinterest style card styling with media queries and remove 3rd party components from HomePage, Feed, and Product components
