+++ frontend/src/components/Feed/Feed.tsx
@@ -2,11 +2,11 @@ import { JSX } from 'react'
 import { useParams } from 'react-router-dom'
 import './Feed.css'
 
-import Col from 'react-bootstrap/Col'
+// import Col from 'react-bootstrap/Col'
 
 import { useGetProductsQuery } from '../../slices/productsApiSlice';
 
-import Product from '@src/components/Product'
+import Product from '@src/components/Product/Product'
 import Loader from '@src/components/Loader'
 import Message from '@src/components/Message'
 
@@ -30,7 +30,7 @@ export default function Feed({ pageNumber }: IFeed): JSX.Element {
       {
         feedData?.products && (
           feedData.products.map((product: IProductKeys): JSX.Element => (
-              <div className="masontry-item" key={product._id} sm={12} md={6} lg={4} xl={3}>
+              <div key={product._id}>
                 <Product product={product} />
               </div>  
             )
