+++ frontend/src/components/Product/Product.css
@@ -1,4 +1,4 @@
-/* .masonry-item {
+.masonry-item {
   display: inline-block;
   width: 100%;
 
@@ -10,10 +10,11 @@
   border-radius: 2px;
 
   background-color: rgba(255, 255, 255, 95%);
- } */
+ }
 
- /* .masonry-item img {
+ .masonry-image {
   display: block;
   width: 100%;
   height: 100%;
- } */
+  /* object-fit: contain; */
+ }
